Added MAF filter support for GDC
